modify to use a flat folder structure
