+++ .eslintrc.js
@@ -1,0 +1,13 @@
+/* eslint-env node */
+require('@rushstack/eslint-patch/modern-module-resolution')
+
+module.exports = {
+	root: true,
+	extends: [
+		'plugin:vue/vue3-essential',
+		'eslint:recommended',
+		'@vue/eslint-config-prettier',
+	],
+	parser: 'vue-eslint-parser',
+	rules: { 'vue/script-setup-uses-vars': 'error' },
+}
