+++ .prettierrc.js
@@ -1,6 +1,6 @@
 module.exports = {
-  trailingComma: 'es5',
-  tabWidth: 2,
-  semi: true,
-  singleQuote: true,
+	trailingComma: 'es5',
+	semi: false,
+	singleQuote: true,
+	useTabs: true,
 }
