+++ package.json
@@ -1,27 +1,34 @@
 {
-  "name": "admin-panel-demo",
-  "version": "0.0.0",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview --port 4173"
-  },
-  "dependencies": {
-    "ant-design-vue": "^3.2.9",
-    "normalize.css": "^8.0.1",
-    "pinia": "^2.0.14",
-    "vue": "^3.2.37",
-    "vue-router": "^4.0.16"
-  },
-  "devDependencies": {
-    "@vitejs/plugin-vue": "^2.3.3",
-    "less": "^4.1.3",
-    "sass": "^1.53.0",
-    "unplugin-vue-components": "^0.20.1",
-    "vite": "^2.9.12",
-    "vite-plugin-pages": "^0.24.2",
-    "vite-plugin-vue-layouts": "^0.6.0",
-    "vite-plugin-windicss": "^1.8.5",
-    "windicss": "^3.5.5"
-  }
+	"name": "admin-panel-demo",
+	"version": "0.0.0",
+	"scripts": {
+		"dev": "vite",
+		"build": "vite build",
+		"preview": "vite preview --port 4173",
+		"lint": "eslint ./src --ext .js --ext .vue"
+	},
+	"dependencies": {
+		"ant-design-vue": "^3.2.9",
+		"normalize.css": "^8.0.1",
+		"pinia": "^2.0.14",
+		"vue": "^3.2.37",
+		"vue-router": "^4.0.16"
+	},
+	"devDependencies": {
+		"@rushstack/eslint-patch": "^1.1.4",
+		"@vitejs/plugin-vue": "^2.3.3",
+		"@vue/eslint-config-prettier": "^7.0.0",
+		"eslint": "^8.19.0",
+		"eslint-plugin-vue": "^9.2.0",
+		"less": "^4.1.3",
+		"prettier": "^2.7.1",
+		"sass": "^1.53.0",
+		"unplugin-vue-components": "^0.20.1",
+		"vite": "^2.9.12",
+		"vite-plugin-pages": "^0.24.2",
+		"vite-plugin-vue-layouts": "^0.6.0",
+		"vite-plugin-windicss": "^1.8.5",
+		"vue-eslint-parser": "^9.0.3",
+		"windicss": "^3.5.5"
+	}
 }
