+++ src/assets/icons/IconHome.vue
@@ -1,26 +1,26 @@
 <template>
-  <svg
-    version="1.1"
-    id="Capa_1"
-    xmlns="http://www.w3.org/2000/svg"
-    xmlns:xlink="http://www.w3.org/1999/xlink"
-    x="0px"
-    y="0px"
-    viewBox="0 0 176.532 176.532"
-    style="enable-background: new 0 0 176.532 176.532"
-    xml:space="preserve"
-  >
-    <g>
-      <path
-        style="fill: #e5e5e5"
-        d="M152.928,85.912l-59.89-49.406c-2.771-2.287-6.773-2.287-9.545,0l-59.89,49.406
+	<svg
+		version="1.1"
+		id="Capa_1"
+		xmlns="http://www.w3.org/2000/svg"
+		xmlns:xlink="http://www.w3.org/1999/xlink"
+		x="0px"
+		y="0px"
+		viewBox="0 0 176.532 176.532"
+		style="enable-background: new 0 0 176.532 176.532"
+		xml:space="preserve"
+	>
+		<g>
+			<path
+				style="fill: #e5e5e5"
+				d="M152.928,85.912l-59.89-49.406c-2.771-2.287-6.773-2.287-9.545,0l-59.89,49.406
 		c-1.728,1.425-2.728,3.546-2.728,5.785v74.544c0,4.143,3.357,7.5,7.5,7.5h119.779c4.143,0,7.5-3.357,7.5-7.5V91.697
 		C155.656,89.458,154.656,87.337,152.928,85.912z M96.196,158.741H80.336v-41.524h15.859V158.741z M140.656,158.741h-29.46v-49.024
 		c0-4.143-3.357-7.5-7.5-7.5H72.836c-4.143,0-7.5,3.357-7.5,7.5v49.024h-29.46V95.233l52.39-43.219l52.39,43.219V158.741z"
-      />
-      <path
-        style="fill: #e5e5e5"
-        d="M173.72,70.866c-16.706-13.382-32.458-26.178-49.561-40.261
+			/>
+			<path
+				style="fill: #e5e5e5"
+				d="M173.72,70.866c-16.706-13.382-32.458-26.178-49.561-40.261
 		c-5.109-4.591-10.479-8.938-15.675-13.144c-5.087-4.118-10.348-8.377-15.201-12.745c-2.854-2.568-7.182-2.568-10.035,0
 		c-4.854,4.368-10.114,8.627-15.201,12.745c-5.195,4.205-10.565,8.553-15.675,13.144C35.27,44.689,19.518,57.484,2.812,70.866
 		c-3.233,2.589-3.755,7.31-1.165,10.542c2.589,3.232,7.311,3.755,10.542,1.165C28.991,69.116,44.829,56.249,62.034,42.08
@@ -28,7 +28,7 @@
 		c3.582,3.046,7.223,5.993,10.781,8.873c5.087,4.118,10.348,8.377,15.201,12.745c0.082,0.074,0.165,0.146,0.25,0.215
 		c17.204,14.169,33.043,27.036,49.845,40.493c1.384,1.108,3.039,1.646,4.684,1.646c2.198,0,4.377-0.962,5.858-2.812
 		C177.475,78.176,176.953,73.455,173.72,70.866z"
-      />
-    </g>
-  </svg>
+			/>
+		</g>
+	</svg>
 </template>
