+++ src/pages/components/data-display.vue
@@ -1,3 +1,3 @@
 <template>
-  <h1>Data Display</h1>
+	<h1>Data Display</h1>
 </template>
