+++ src/pages/components/data-entry.vue
@@ -1,3 +1,3 @@
 <template>
-  <h1>Data Entry</h1>
+	<h1>Data Entry</h1>
 </template>
