+++ src/pages/components/feedback.vue
@@ -1,3 +1,3 @@
 <template>
-  <h1>Feedback</h1>
+	<h1>Feedback</h1>
 </template>
