+++ src/pages/components/general.vue
@@ -1,3 +1,3 @@
 <template>
-  <h1>General</h1>
+	<h1>General</h1>
 </template>
