+++ src/pages/components/navigation.vue
@@ -1,3 +1,3 @@
 <template>
-  <h1>Navigation</h1>
+	<h1>Navigation</h1>
 </template>
